refactor(cli): collapse app/main.py to entry-point shim
Final step of the cli/ extraction. app/main.py is now a 5-line shim
that re-exports `app` from app.cli for the `atelier = app.main:app`
entry point. The two render helpers that test_main.py was importing
through app.main (_render_task_event, _truncate_tail) are no longer
re-exported there; the tests now import them from their canonical
home in app.cli.render.

Checkpoint C verified:
  - app/main.py is 5 lines (≤ 10 required).
  - 361 unit tests pass; 2 pre-existing live-harness failures excluded.
  - Every --help is byte-identical to tasks/baseline_help/.
  - No file imports from app.main anywhere in app/ or tests/.
  - cli/render.py at 274 lines slightly exceeds the 250-line rough
    sanity bound; documented in tasks/todo.md as out-of-scope cleanup.
