+++ app/main.py
@@ -1,9 +1,5 @@
-"""Typer CLI entrypoint for flow-atelier."""
-from __future__ import annotations
-
+"""Backwards-compat entry point for `atelier = app.main:app`."""
 from app.cli import app
-from app.cli.render import _render_task_event, _truncate_tail  # noqa: F401
-
 
 if __name__ == "__main__":
     app()
